+++ internal/transport/grpc/interceptors/recovery.go
@@ -1,0 +1,24 @@
+package interceptors
+
+import (
+	"context"
+
+	"github.com/CrispyCl/TestGolangUsers/pkg/logger"
+	"github.com/grpc-ecosystem/go-grpc-middleware/v2/interceptors/recovery"
+	"go.uber.org/zap"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+)
+
+func RecoveryInterceptor(ctx context.Context, log logger.Logger) grpc.UnaryServerInterceptor {
+	recoveryOpts := []recovery.Option{
+		recovery.WithRecoveryHandler(func(p any) (err error) {
+			log.Error(ctx, "Recovered from panic", zap.Error(err))
+
+			return status.Errorf(codes.Internal, "internal error")
+		}),
+	}
+
+	return recovery.UnaryServerInterceptor(recoveryOpts...)
+}
